requirements.txt
langchain==0.0.254
openai==0.27.8
pypdf2==3.0.1
python-dotenv==1.0.0
qdrant-client==1.3.1
streamlit==1.24.1
tiktoken==0.4.0